% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ACS_variables_function_new.R
\docType{data}
\name{acs_years}
\alias{acs_years}
\title{Function to retrieve the available ACS demographic variables and their codes for a specified survey year}
\format{
An object of class \code{character} of length 3.
}
\usage{
acs_years
}
\arguments{
\item{acs_year}{(string). The desired 5-Year ACS year (ex: for the 2017-2021 5-Year ACS, enter '2021')}
}
\value{
dataframe of available variables
}
\description{
Function to retrieve the available ACS demographic variables and their codes for a specified survey year
}
\keyword{datasets}
