% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ACS_variables_function.R
\name{get_ACS_variables}
\alias{get_ACS_variables}
\title{Function to retrieve the available ACS demographic variables and their codes for a specified survey year}
\usage{
get_ACS_variables(acs_year = NULL)
}
\arguments{
\item{acs_year}{(string). The desired 5-Year ACS year (ex: for the 2017-2021 5-Year ACS, enter "2021"). The most recent year available will populate if left NULL.}
}
\value{
Table of available variables with columns for variable code, variable name, denominator code, and denominator name (the "denominator variable" is the denominator population in percent estimate calculations). Use variable codes as the input for `var_codes` in get_geo_estimates().
}
\description{
Function to retrieve the available ACS demographic variables and their codes for a specified survey year
}
