% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bbl_estimates_function.R
\name{get_bbl_estimates}
\alias{get_bbl_estimates}
\title{Function to produce a dataframe containing BBL-level population estimates for a specified year}
\usage{
get_bbl_estimates(year = NULL)
}
\arguments{
\item{year}{(string). The desired year for BBL-level estimates. The most recent year available will populate if left NULL.}
}
\value{
Table with population estimates by BBL, as well as multiple geography columns for aggregation. Avoid using estimates for individual BBLs—the more aggregation, the less error.
}
\description{
Function to produce a dataframe containing BBL-level population estimates for a specified year
}
