% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bbl_estimates_function.R,
%   R/get_bbl_estimates_function_new.R
\name{get_bbl_estimates}
\alias{get_bbl_estimates}
\title{Function to retrieve PLUTO bbl population estimates}
\usage{
get_bbl_estimates(year)

get_bbl_estimates(year)
}
\arguments{
\item{year}{(string). The desired year for BBL-level estimates}
}
\value{
tibble with PLUTO bbl population estimates

tibble with PLUTO bbl population estimates
}
\description{
Function to retrieve PLUTO bbl population estimates

A dataframe containing BBL-level population estimates for a specified year
}
